The campaign's MossTank (VTank scheduler, key-holding mover, door rule,
fellow vitals, 9/9 live proof) meets main's shared host: the Runtime
automation surface from #121/#130, the collision-built navigation from
#131, the plugin API documentation. Resolution rule: host code takes
main's implementation; the plugin keeps the campaign's; the contract keeps
both sides' additions.

What the campaign added to the host and main lacked is ported onto main's
surface rather than kept as a second implementation: chat input-active and
composer bindings, unbuffed base vitals, container Close for items and
loot, the fellowship vitals request and vitals age, the weapon fields on
owned equipment, item effects, corpse positions; the scoped plugin chat
forwards the two new chat members. Hostile-target queries pass the scope
the campaign introduced.

Dropped here, to come back in the new shape: the "walk legs with client
pathing" checkbox and its per-leg GoTo walk, whose replacement (stuck-only
hand-off, three-way setting, pause on a failed leg) is the next commits.
The campaign's App/Headless parity script goes with the App-side surface it
compared. Chat log-text types are the int main chose.

Suites (Release, CI lane filter): Core 5373, Runtime 2285, Headless 396,
App 8205, MossTank 1350, all passing.

+++ src/AcDream.Plugins.MossTank/CombatController.cs
@@ -2653,7 +2653,7 @@ internal sealed class CombatController
                     && message.SenderObjectId != 0u
                     && message.SenderObjectId
                         == _host.Automation.Character.ObjectId,
-                logTextType: message.LogTextType);
+                logTextType: (uint)message.LogTextType);
         }
     }
